Further fixed some code

+++ src/main/java/vn/giakhanhvn/skysim/entity/caverns/LapisZombie.java
@@ -39,7 +39,7 @@ EntityStatistics {
 
     @Override
     public SEntityEquipment getEntityEquipment() {
-        return new SEntityEquipment(null, new ItemStack(Material.STAINED_GLASS, 1, 11), SItem.of(SMaterial.LAPIS_ARMOR_CHESTPLATE).getStack(), SItem.of(SMaterial.LAPIS_ARMOR_LEGGINGS).getStack(), SItem.of(SMaterial.LAPIS_ARMOR_BOOTS).getStack());
+        return new SEntityEquipment(null, new ItemStack(Material.STAINED_GLASS, 1, (short) 11), SItem.of(SMaterial.LAPIS_ARMOR_CHESTPLATE).getStack(), SItem.of(SMaterial.LAPIS_ARMOR_LEGGINGS).getStack(), SItem.of(SMaterial.LAPIS_ARMOR_BOOTS).getStack());
     }
 
     @Override
